requirements.txt
Django==3.0.5
djangorestframework==3.11.0
django-cors-headers==3.2.1
psycopg2==2.8.5
daphne==2.5.0
django-rest-auth==0.9.5
django-allauth==0.42.0